finding tempetarure for all capitals
